:truck: refactor: 파일의 위치를 옮긴다
- ViewModelFactory 를 각 패키지로 이동한다
