+++ app/src/main/java/woowacourse/shopping/cart/CartItemRecyclerViewAdapter.kt
@@ -3,8 +3,8 @@ package woowacourse.shopping.cart
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.data.Product
 import woowacourse.shopping.databinding.HolderCartBinding
-import woowacourse.shopping.db.Product
 
 class CartItemRecyclerViewAdapter(
     private var values: List<Product>,
