+++ app/src/main/java/woowacourse/shopping/productList/ProductListFragment.kt
@@ -9,7 +9,6 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.ProductListViewModelFactory
 import woowacourse.shopping.R
 import woowacourse.shopping.TwentyItemsPagingStrategy
 import woowacourse.shopping.cart.ShoppingCartFragment
