+++ app/src/main/java/woowacourse/shopping/repository/DummyShoppingCartItemRepository.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.repository
 
 import woowacourse.shopping.PagingStrategy
-import woowacourse.shopping.db.Product
+import woowacourse.shopping.data.Product
 
 class DummyShoppingCartItemRepository(private val pagingStrategy: PagingStrategy<Product>) :
     ShoppingCartItemRepository {
