+++ app/src/main/res/layout/holder_product.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="product"
-            type="woowacourse.shopping.db.Product" />
+            type="woowacourse.shopping.data.Product" />
 
     </data>
